+++ lib/data_seeder/config.rb
@@ -1,11 +1,12 @@
 module DataSeeder
   class Config
-    attr_accessor :seed_dir, :logger, :loaders
+    attr_accessor :seed_dirs, :logger, :loaders
 
     def initialize
-      @seed_dir = 'db/seed'
-      @logger   = Logger.new
-      @loaders  = default_loaders
+      @seed_dirs  = ['db/seed'].freeze
+      @logger     = Logger.new
+      @loaders    = default_loaders
+      @is_default = true
     end
 
     def verbose=(verbose)
@@ -36,6 +37,22 @@ module DataSeeder
 
     def add_loader(ext, loader)
       @loaders[ext] = loader
+    end
+
+    def seed_dir=(seed_dir)
+      @seed_dirs = [seed_dir]
+    end
+
+    def seed_dir
+      @seed_dirs.first
+    end
+
+    def add_seed_dir(seed_dir)
+      if @seed_dirs.frozen?
+        @seed_dirs = [seed_dir]
+      else
+        @seed_dirs << seed_dir
+      end
     end
   end
 end