+++ lib/data_seeder/loader.rb
@@ -2,68 +2,69 @@ require 'English'
 
 module DataSeeder
   module Loader
-    attr_accessor :file_config, :key_attribute
-    attr_reader   :path, :path_minus_ext
-
-    def initialize(options={})
-      @only = options[:only]
-      @except = options[:except]
-      if options.has_key?(:purge)
-        @purge = options[:purge]
-      else
-        @purge = true
-      end
-      @old_keys = []
-    end
-
-    def config
-      DataSeeder.config
-    end
+    attr_accessor :key_attribute
+    attr_reader   :klass, :path, :path_minus_ext, :config
 
     def logger
       DataSeeder.logger
     end
 
-    def klass!
-      @path_minus_ext.classify.constantize
-    rescue NameError => e
-      raise "#{@path} doesn't match a corresponding model"
+    def default_config
+      { purge: true }
     end
 
-    def klass
-      # This should always translate to a class except for custom loaders
-      @path_minus_ext.classify.constantize rescue nil
-    end
-
+    # Returns nil if the path was not ready for processing due to dependencies
     def process(path)
       @path           = path
       dot_index       = @path.rindex('.')
       @path_minus_ext = @path[0, dot_index]
-      @file_config    = {}
       cfg_file = "#{@path_minus_ext}.cfg"
-      @file_config = eval(File.read(cfg_file)) if File.exist?(cfg_file)
-      File.open(@path, 'r') do |fin|
-        load_file_config(fin) if @file_config.empty?
-        @file_config = ActiveSupport::HashWithIndifferentAccess.new(@file_config)
-        setup
-        load(fin)
-        teardown
+      begin
+        @klass = @path_minus_ext.classify.constantize
+      rescue NameError => e
+        @klass = nil
       end
-      call_file_method(:teardown)
+      @config = default_config
+      if @klass && @klass.respond_to?(:data_seeder_config)
+        @config = @config.merge(@klass.data_seeder_config)
+      end
+      if File.exist?(cfg_file)
+        @config = @config.merge(eval(File.read(cfg_file)))
+      end
+      File.open(@path, 'r') do |io|
+        # TODO: Lets get rid of this load_file_config stuff.  It's too hacky and may cause problems for
+        # non-file IO types.
+        load_file_config(io)
+        depends = call_method(:depends)
+        return nil if depends && !SeedFile.processed?(depends)
+        process_io(io)
+      end
+      call_method(:teardown)
+      return true
+    end
+
+    def process_io(io, config={})
+      @config = (@config || default_config).merge(config)
+      @key_attribute = @config[:key_attribute] || :id
+      setup
+      load(io)
+      teardown
     end
 
     def setup
-      @key_attribute = self.file_config[:key_attribute] || :id
-      @old_keys = self.klass!.all.pluck(@key_attribute).map(&:to_s) if @purge
+      if @config[:purge]
+        @old_ids = klass.all.pluck(:id).to_set
+      else
+        @old_ids = Set.new
+      end
       logger.info { "Loading #{@path}" }
-      call_file_method(:setup)
+      call_method(:setup)
     end
 
     def teardown
-      @old_keys.each do |key|
-        if model = self.klass!.find_by(@key_attribute => key)
-          logger.info { "  Destroying #{model_info(model)}"}
-          model.destroy
+      @old_ids.each do |id|
+        if model = klass.find_by(id: id)
+          destroy_model(model)
         end
       end
     end
@@ -72,26 +73,23 @@ module DataSeeder
     # The changes argument will be the model.changes on an update.
     def model_info(model, changes=nil)
       if changes
-        attr = @file_config[:update_display_method] || @key_attribute
+        attr = @config[:update_display_method] || @key_attribute
         "#{model.send(attr)}: #{changes.inspect}"
       else
         model.inspect
       end
     end
 
-    def load_file_config(fin)
-      config_line = fin.readline
+    def load_file_config(io)
+      config_line = io.readline
       if match = config_line.match(/^\s*#\s*config:(.*)/)
-        @file_config = eval(match[1])
+        @config = @config.merge(eval(match[1]))
       else
-        fin.seek(0)
-        if self.klass && self.klass.respond_to?(:data_seeder_config)
-          @file_config = self.klass.data_seeder_config
-        end
+        io.seek(0)
       end
     end
 
-    def load(fin)
+    def load(io)
       throw 'Must override load'
     end
 
@@ -100,40 +98,62 @@ module DataSeeder
     end
 
     def save(attr)
-      if @file_config[:use_line_number_as_id]
+      attr = call_method(:preprocess, attr) || attr
+      if @config[:use_line_number_as_id]
         key = self.line_number
       else
         key = attr[@key_attribute.to_s] || attr[@key_attribute.to_sym]
         raise "No #{@key_attribute} in #{attr.inspect}" unless key
       end
-      if method = @file_config[:postprocess]
-        method.call(attr)
-      elsif self.klass!.respond_to?(:data_seeder_postprocess)
-        self.klass!.send(:data_seeder_postprocess, attr)
-      end
-      @old_keys.delete(key.to_s)
-      model = self.klass!.find_or_initialize_by(@key_attribute => key)
+      call_method(:postprocess)
+      model = self.klass.find_or_initialize_by(@key_attribute => key)
       model.attributes = attr
       save_model(model)
     end
 
     def save_model(model)
       if model.new_record?
-        logger.info { "  Saving #{model_info(model)}" }
+        log_save(model)
       else
-        changes = model.changes
-        return if changes.empty?
-        logger.info { "  Updating #{model_info(model, changes)}" }
+        @old_ids.delete(model.id)
+        return unless model.changed?
+        log_update(model)
       end
       model.save!
     end
 
-    def call_file_method(name, *args)
-      if method = @file_config[name]
-        return method.call(*args)
+    # Allow override for potential soft-delete
+    def destroy_model(model)
+      log_destroy(model)
+      model.destroy
+    end
+
+    def log_save(model)
+      #logger.info { "  Saving #{model_info(model)}" }
+    end
+
+    def log_update(model)
+      logger.info { "  Updating #{model_info(model, model.changes)}" }
+    end
+
+    def log_destroy(model)
+      logger.info { "  Destroying #{model_info(model)}"}
+    end
+
+    def call_method(name, *args)
+      if ![:setup,:teardown].include?(name) && self.respond_to?(name)
+        return send(name, @args)
+      elsif val = @config[name]
+        if val.kind_of?(Proc)
+          return val.call(*args)
+        else
+          return val
+        end
       else
-        class_method = "data_seeder_#{name}"
-        return self.klass.send(class_method, *args) if @klass.respond_to?(class_method)
+        if @klass
+          class_method = "data_seeder_#{name}"
+          return @klass.send(class_method, *args) if @klass.respond_to?(class_method)
+        end
       end
       return nil
     end